+++ bbinc/roll_file.h
@@ -1,6 +1,6 @@
 #ifndef INCLUDED_ROLLFILE_H
 #define INCLUDED_ROLLFILE_H
 
-int roll_file(char *file, int nkeep);
+int roll_file(const char *file, int nkeep);
 
 #endif
